requirements.txt
django~=4.2.13
psycopg2-binary~=2.9.9
djangorestframework~=3.15.2
python-dotenv~=1.0.1
pillow~=10.3.0
django-filter==24.2
djangorestframework-simplejwt~=5.3.1
coverage==7.6.0
celery~=5.4.0
eventlet~=0.36.1
django-celery-beat~=2.6.0
pytz~=2024.1